correcion de comit anterior, (no lo guarde bien)

+++ Lab01_SJH/Models/Player.cs
@@ -4,7 +4,7 @@ namespace Lab01_SJH.Models
     public class Player : Coach
     {
         public string Rol { get; set; }
-        public decimal KDA { get; set; }
+        public double KDA { get; set; }
         public int CS { get; set; }
     }
 }
